Fix type inconsistencies in ParameterBlakcboard API
The recently-introduced methods readBytes() and writeBytes() where using a
uint32_t for the 'offset' argument whereas all other methods use a size_t. This
is inconsistent and also will trigger a conversion warning with -Wconversion.

+++ parameter/ParameterBlackboard.cpp
@@ -95,14 +95,14 @@ void CParameterBlackboard::readBuffer(void* pvDstData, size_t size, size_t offse
 }
 
 // Element access
-void CParameterBlackboard::writeBytes(const std::vector<uint8_t>& bytes, uint32_t offset)
+void CParameterBlackboard::writeBytes(const std::vector<uint8_t>& bytes, size_t offset)
 {
     assertValidAccess(offset, bytes.size());
 
     std::copy(begin(bytes), end(bytes), atOffset(offset));
 }
 
-void CParameterBlackboard::readBytes(std::vector<uint8_t>& bytes, uint32_t offset) const
+void CParameterBlackboard::readBytes(std::vector<uint8_t>& bytes, size_t offset) const
 {
     assertValidAccess(offset, bytes.size());
